Move dropped connection handler to db module
