+++ swift_x_account_sharing/auth.py
@@ -15,6 +15,9 @@ import time
 import logging
 
 import aiohttp.web
+from asyncpg import InterfaceError
+
+from .db import handle_dropped_connection
 
 
 AiohttpHandler = typing.Callable[
@@ -96,10 +99,13 @@ async def handle_validate_authentication(
                 project = None
     finally:
         if project:
-            project_tokens = [
-                rec["token"].encode("utf-8")
-                for rec in await request.app["db_conn"].get_tokens(project)
-            ]
+            try:
+                project_tokens = [
+                    rec["token"].encode("utf-8")
+                    for rec in await request.app["db_conn"].get_tokens(project)
+                ]
+            except InterfaceError:
+                handle_dropped_connection(request)
         else:
             LOGGER.debug(f"No project ID found in request {request}")
             if request.path != "/health":
