+++ swift_x_account_sharing/db.py
@@ -8,6 +8,27 @@ import random
 import typing
 
 import asyncpg
+import aiohttp.web
+
+
+MODULE_LOGGER = logging.getLogger("db")
+
+
+def handle_dropped_connection(
+        request: aiohttp.web.Request
+):
+    """Handle dropped database connection."""
+    MODULE_LOGGER.log(
+        logging.ERROR,
+        "Lost database connection, reconnecting..."
+    )
+    request.app["db_conn"].erase()
+    asyncio.ensure_future(
+        request.app["db_conn"].open()
+    )
+    raise aiohttp.web.HTTPServiceUnavailable(
+        reason="No database connection."
+    )
 
 
 class DBConn:
@@ -16,7 +37,7 @@ class DBConn:
     def __init__(self):
         """Initialize connection variable."""
         self.conn = None
-        self.log = logging.getLogger("db")
+        self.log = MODULE_LOGGER
 
     def erase(self):
         """Erase the connection."""
